fix: list nearby posts using meters instead of kilometers

+++ src/main/java/com/locarie/backend/controllers/PostController.java
@@ -4,11 +4,12 @@ import com.locarie.backend.domain.dto.PostDto;
 import com.locarie.backend.services.impl.post.PostServiceImpl;
 import com.locarie.backend.services.post.PostService;
 import jakarta.validation.Valid;
-import java.util.List;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+
+import java.util.List;
 
 @RestController
 @RequestMapping("/api/v1/posts")
@@ -36,7 +37,7 @@ public class PostController {
   public List<PostDto> listNearby(
       @RequestParam(value = "latitude") double latitude,
       @RequestParam(value = "longitude") double longitude,
-      @RequestParam(value = "distance") int distance) {
+      @RequestParam(value = "distance") double distance) {
     return service.listNearby(latitude, longitude, distance);
   }
