+++ src/main/java/com/locarie/backend/services/impl/post/PostServiceImpl.java
@@ -4,11 +4,12 @@ import com.locarie.backend.domain.dto.PostDto;
 import com.locarie.backend.services.post.PostCreateService;
 import com.locarie.backend.services.post.PostReadService;
 import com.locarie.backend.services.post.PostService;
-import java.util.List;
-import java.util.Optional;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
+
+import java.util.List;
+import java.util.Optional;
 
 @Service
 public class PostServiceImpl implements PostService {
@@ -38,7 +39,7 @@ public class PostServiceImpl implements PostService {
   }
 
   @Override
-  public List<PostDto> listNearby(double latitude, double longitude, int distance) {
+  public List<PostDto> listNearby(double latitude, double longitude, double distance) {
     return readService.listNearby(latitude, longitude, distance);
   }
 }
